+++ CHANGELOG.md
@@ -9,8 +9,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 
 ## 2.0.0 - 2025-10-13
+### Added
 - Added `defineAdminBarElements` method that makes it easier to define Admin Bar Component’s custom elements.
   - Calling this method defines the `<admin-bar>` and any other items specified. Sending in a blank array will still define `<admin-bar>`.
+
+### Removed
+- Removed `--admin-bar-gradient-direction` variable
 
 
 ## 1.10.1 - 2025-10-08